Refactor backend to standard hardhat example
