+++ .gitignore
@@ -29,3 +29,9 @@ dist-ssr
 
 # vite-bundle-visualizer output
 /frontend/stats.html
+
+# Ignore backend autogenerated files
+backend/abis
+backend/artifacts
+backend/cache
+backend/src
